+++ src/components/Header/Header.js
@@ -2,7 +2,7 @@ import { Link } from "gatsby"
 import PropTypes from "prop-types"
 import React from "react"
 
-import SocialLinks from "./SocialLinks"
+import SocialLinks from "../SocialLinks/SocialLinks"
 import "./Header.css"
 
 const Header = ({ siteTitle }) => (
